% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadir_bc.R
\name{datadir_bc}
\alias{datadir_bc}
\title{Set local directory for storage of raster data files}
\usage{
datadir_bc(select = FALSE, data.dir = NULL)
}
\arguments{
\item{select}{A boolean indicating whether the directory should be automatically selected (default) or user-selected}

\item{data.dir}{An (optional) character string, providing the absolute path to the desired storage directory}
}
\value{
character string, the absolute path to the selected storage directory
}
\description{
All files downloaded/created by the rasterbc package are written to the directory assigned by this function. The path to
this directory is stored in the global options list for R under 'rasterbc.data.dir'
}
\details{
By default this is the path returned by \code{rappdirs::user_data_dir}. With argument \code{select==TRUE}, a different
directory can be specified: If this is not specified in \code{data.dir}, the user is prompted to choose it via an interactive
file dialog.
}
\examples{
datadir_bc(TRUE, 'H:/')
}
