% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_bc.R
\name{getdata_bc}
\alias{getdata_bc}
\title{Download multiple blocks, and (optionally) merge/clip/mask them}
\usage{
getdata_bc(
  geo = NULL,
  collection = NULL,
  varname = NULL,
  year = NULL,
  force.dl = FALSE
)
}
\arguments{
\item{geo}{vector of character strings (NTS/SNRC codes) or any geometry set of type sfc}

\item{collection}{character string, indicating the data collection to query}

\item{varname}{character string, indicating the layer to query}

\item{year}{integer, indicating the year to query}

\item{force.dl}{boolean, indicates to download (and overwrite) any existing data}
}
\value{
A character string (or vector of them) containing the absolute path(s) to the file(s) written
}
\description{
Downloads and/or loads all blocks covering the geographical extent specified in input argument `geo`. This can be a vector of
(4-character) NTS/SNRC block codes, or any geometry set of type `sfc`.
}
\details{
Data for the layer specified by `collection`, `varname`,
and `year` are written to the directory specified by `rasterbc::datadir_bc`, and blocks already downloaded are not downloaded
again (unless `force.dl==TRUE`).

Note that if no extent object (`geo`) is supplied, then all required blocks are downloaded. Missing arguments to `collection`,
`varname`, and `year` will prompt the function to download ALL corresponding layers. eg. getdata_bc('dem') will
download all three layers (`dem`, `aspect`, `slope`) in the `dem` collection; and getdata_bc('fids', year=2012) will download
'fids' layers from the year 2012.
}
