% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_bc}
\alias{metadata_bc}
\title{Metadata for rasterbc collections}
\format{
Nested list containing metadata and URLs for all of the files available through rasterbc
\describe{
  \item{dem}{Digital Elevation Model data from Natural Resources Canada}
}
}
\source{
Various Canadian government environment ministries
}
\usage{
metadata_bc
}
\description{
This will soon be filled in with a shortened version of the info available at the
\href{https://github.com/deankoch/rasterbc_src}{rasterbc_src} github repository. The script ('metadata.R')
used to generate this file can be found in the rasterbc/data-raw/ subdirectory.
}
\keyword{datasets}
