% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NTS.polygons}
\alias{NTS.polygons}
\title{National Topographic System Index Maps for British Columbia.}
\format{
Simple feature collection with 89 features and 1 field:
\describe{
  \item{NTS_SNRC}{four-character mapsheet code}
}
}
\source{
\url{https://www.nrcan.gc.ca/earth-sciences/geography/topographic-information/maps/9765}
}
\usage{
NTS.polygons
}
\description{
Polygons delineating mapsheet boundaries for 1:250,000 maps, from Natural Resources Canada
}
\keyword{datasets}
