% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadir_bc.R
\name{datadir_bc}
\alias{datadir_bc}
\title{Set local directory for storage of raster data files}
\usage{
datadir_bc(data.dir = NULL, suppress.warning = FALSE)
}
\arguments{
\item{data.dir}{A character string, providing the absolute path to the desired storage directory}

\item{suppress.warning}{A boolean indicating whether to warn if selecting directory that already contains files}
}
\value{
character string, the absolute path to the selected storage directory
}
\description{
All files downloaded/created by the rasterbc package are written to the directory assigned by this function. The path to
this directory is stored in the global options list for R under 'rasterbc.data.dir'
}
\details{
By default this is the path returned by \code{rappdirs::user_data_dir}. The optional argument \code{data.dir} specifies
an alternative directory manually as a character string (absolute file path). The directory will be created if it doesn't already
exist. Its path is stored as an R option: to check the current setting for data.dir, run \code{getOption('rasterbc.data.dir')}
}
\examples{
datadir_bc()
}
