% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findblocks_bc.R
\name{findblocks_bc}
\alias{findblocks_bc}
\title{Identify NTS/SNRC codes covering a given geometry}
\usage{
findblocks_bc(input.sf = NULL)
}
\arguments{
\item{input.sf}{A simple features object (class sfc), at least part of which intersects with the province of BC}
}
\value{
A vector of character strings, the 4-character NTS/SNRC codes of the mapsheets that overlap with input.sf
}
\description{
Data layers for the province of BC are split into blocks, according to the mapsheet coding system used by Canada's NTS/SNRC.
This function identifies which blocks are needed to cover a given (user-supplied) geographical extent.
}
\details{
Returns the 4-character NTS/SNRC codes corresponding to mapsheets that intersect with the geometry set(s) in input.sf. If input.sf
is NULL, returns the character vector of all codes, ordered according to the rows of 'ntspoly_bc'.
}
\examples{
findblocks_bc()
}
