% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listdata_bc.R
\name{listdata_bc}
\alias{listdata_bc}
\title{List all available layers, or their associated filepaths and download status}
\usage{
listdata_bc(
  collection = NULL,
  varname = NULL,
  year = NULL,
  verbose = 1,
  return.boolean = FALSE
)
}
\arguments{
\item{collection}{(Optional) character string, indicating the data collection to query}

\item{varname}{(Optional) character string, indicating the layer to query (see Details, below)}

\item{year}{(Optional) integer, indicating the year to query (see Details, below)}

\item{verbose}{An integer (0, 1, 2), indicating how much information about the files to print to the console}

\item{return.boolean}{A boolean, indicating whether to return a dataframe or a nested list of booleans}
}
\value{
Either a (list of) dataframe(s) containing information about each raster layer, or (when `return.boolean==TRUE`) a
nested list of boolean values (named according to filepath), with entries for each data file in the specified subset.
}
\description{
Returns a list of dataframes (one per collection) describing the variables available through rasterbc, or a subset (as
specified by `collection`, `varname`, `year`). Alternatively, if `return.boolean==TRUE`, returns a nested list indicating
the filepaths associated with these layers, and which of them exist in the local data storage directory already.
}
\details{
The layers available through this package are organized into "collections", corresponding to their original online sources.
Layers in a collection are further organized by variable name, and are uniquely identified by the character string `varname`
(and, if applicable, `year`). The optional arguments `collection`, `varname`, `year` prompt this function to return only the
applicable subsets.
}
\examples{
x = listdata_bc()
x = listdata_bc(collection='bgcz', verbose=1)
x = listdata_bc(collection='dem', verbose=2)
x = listdata_bc(collection='dem', varname='aspect', verbose=2)
x = listdata_bc(collection='fids', varname='IBM_trace', year=2005, verbose=2)
print(x)

}
