% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listdata_bc.R
\name{listdata_bc}
\alias{listdata_bc}
\title{Identify which data layers are available online, and (optionally) which have been downloaded already}
\usage{
listdata_bc(collection = NULL, varname = NULL, year = NULL, verbose = 1)
}
\arguments{
\item{collection}{character string, indicating the data collection to query}

\item{varname}{character string, indicating the layer to query (see Details, below)}

\item{year}{integer, indicating the year to query (see Details, below)}

\item{verbose}{integer (0, 1, 2), indicating how much information about the files to print to the console}
}
\value{
A (named) nested list of boolean values, one for each file. This list has the same structure and naming scheme as
'rasterbc::metadata_bc', but with boolean entries instead of character strings. The entry indicates whether the file exists on
disk in the local data storage directory specified by 'rasterbc::datadir_bc'.
}
\description{
Prints a list of collection and variable names available through rasterbc, and (optionally) returns a nested list indicating
which of these exist on disk already
}
\details{
The layers available through this package are organized into "collections", corresponding to their original online source.
Layers in a collection are further organized by variable name, and are uniquely identified by the character string "varname"
(and, if applicable, the year). To see the full list of variable names by collection, run

When arguments collection and/or varname are provided, check only for those (sub)collections, returning
the corresponding (sub)list. This saves having to scan the entire data storage directory, which can be slow.
}
\examples{
x = listdata_bc()
x = listdata_bc(collection='bgcz', verbose=1)
x = listdata_bc(collection='dem', verbose=2)
x = listdata_bc(collection='dem', varname='aspect', verbose=2)
x = listdata_bc(collection='fids', varname='IBM_trace', year=2005, verbose=2)
print(x)

}
