% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendata_bc.R
\name{opendata_bc}
\alias{opendata_bc}
\title{Load blocks and (optionally) merge/clip/mask them}
\usage{
opendata_bc(
  geo = NULL,
  collection = NULL,
  varname = NULL,
  year = NULL,
  load.mode = "mask"
)
}
\arguments{
\item{geo}{vector of character strings (NTS/SNRC codes) or any geometry set of type sfc}

\item{collection}{character string, indicating the data collection to query}

\item{varname}{character string, indicating the layer to query}

\item{year}{integer, indicating the year to query}

\item{load.mode}{character string, one of 'all', 'clip', 'mask'}
}
\value{
A `rasterLayer` (or list of them)
}
\description{
Loads (as a single RasterLayer) all blocks covering the geographical extent specified in input argument `geo`. This can be a vector of
(4-character) NTS/SNRC block codes, or any geometry set of type `sfc`.
}
\details{
Data for the layer specified by `collection`, `varname`, and (as needed) `year`, are fetched from the directory specified by
\code{\link{datadir_bc}}, merged into a single (mosaic) layer, cropped and masked as needed, and then loaded into memory and returned
as a \code{\link{RasterLayer-class}} object.

Argument `load.mode` specifies whether/how the blocks should be loaded into memory. 'all' uses `gdalUtils::gdal_mosaic` to merge the
blocks into a larger rasterLayer, covering all (and likely more) of the input extent; 'clip' and 'mask' (the default) assume that `geo`
is a (multi)polygon, cropping the merged blocks to its boundary; and  with 'mask', all data not inside the polygon is set to NA.
}
